Rewrite build scripts
This adds the new dependencies for the latest dl-fldigi and updates
some old ones. There are disadvantages and advantages to doing this in
python vs a large makefile: python isn't really so suited to lots of
shell commands, whereas as implementing the state.json and having the
verbosity settings and logging would be completely disgusting in a makefile.
Overall this works out okay.

Usage: ./builder.py git-location [git-commit]
Use --help for help. It has several options.

Dependencies will be built in a build/ directory. Each dependency is
installed under its own prefix, and the source is removed afterwards.
The script will setup this directory the first time it is run, and then
clean up. It will come to about 40MB. Then, on the first and every time
after, it will build dl-fldigi.

Downloads are md5summed and cached in a cache folder. This folder could
be removed after the initial dependency build.

Warning! Only tested on Ubuntu lucid (10.04) so far.